Adds AsyncTestCase and TestContext
- Adds an AsyncTestCase that allows you to easily run tests on amphp's
Loop and ensure that your test runs to completion by returning a Promise
or a Generator that controls when the test finishes. There is also a
"safety valve" that ensures tests do not run past a certain period of
time; by default this value is 1500ms but can be changed in your TestCases
by overriding the AsyncTestCase::getTestTimeout() method.

- Adds a TestContext object that helps ensure your tests needing to run
multiple async tasks are able to run to completion by allowing you to
require a number of TestContext::checkpoint() calls to resolve a
Promise associated to the TestContext. This class is intended to be used
in conjunction with the AsyncTestCase.