Add AsyncTestCase (#2)
Adds an AsyncTestCase that allows to easily run tests on Amp's event loop and ensure that the tests runs to completion by returning a Promise or Generator that controls when the test finishes. There is also a "safety valve" that ensures tests do not run past a certain period of time.